Add HttpMethod interface and Standard HttpMethods
A new interface for HTTP Methods has been created, with means to retrieve the
method name and whether if it is idempotent and/or safe. Another method to
match on HttpExchange objects has also been default-implemented in this
interface.

StandardHttpMethods is an enum implementing this HttpMethod API for al standard
methods defined in https://www.iana.org/assignments/http-methods/http-methods.xhtml.