Implement controller and view for /login GET request
A new LoginView has been included along a LoginController for correctly
rendering it when the server receives the adequate request. New Route and
Router classes have also been implemented in order to ease the creation of
method/path (and any other filter) routes and association to their respective
controllers.

New Main class has also been created to manually test that the developed
infrastructure and this new LoginView is working as expected.